Fix integer overflow problem in framerate calculation (that occurs on my slow pc...) Another option would be measure time in milliseconds, which should be accurate enough for this purpose, rather than nanoseconds?
(from subversion r5746)

+++ greenfoot/src/greenfoot/core/Simulation.java
@@ -257,7 +257,7 @@ public class Simulation extends Thread
                 lastRepaintTime = repaintTimes.peek();
             }
         }
-        int timeSinceRepaint = (int) (currentTime - lastRepaintTime);
+        long timeSinceRepaint = currentTime - lastRepaintTime;
         // Avoid divide by zero
         if (timeSinceRepaint == 0)
             timeSinceRepaint = 1;
